Revert "feat: add support for differential zip updates on macOS (#7709)"
This reverts commit 79df54238621fbe48ba20444129950ba2dc49983.
